+++ app/src/main/res/menu/notes_list_main_menu.xml
@@ -7,18 +7,6 @@
         android:title="@string/settings"
         app:showAsAction="never" />
     <item
-        android:id="@+id/action_main"
-        android:icon="@android:drawable/ic_menu_compass"
-        android:orderInCategory="10"
-        android:title="@string/main"
-        app:showAsAction="always" />
-    <item
-        android:id="@+id/action_favorite"
-        android:icon="@android:drawable/btn_star_big_on"
-        android:orderInCategory="20"
-        android:title="@string/favorite"
-        app:showAsAction="ifRoom" />
-    <item
         android:id="@+id/action_search"
         android:icon="@android:drawable/ic_menu_search"
         android:orderInCategory="5"
