+++ app/src/main/res/values/strings.xml
@@ -16,4 +16,5 @@
     <string name="main">Main action</string>
     <string name="favorite">Favorite</string>
     <string name="search">Search</string>
+    <string name="share">Share</string>
 </resources>
